requirements.txt
altgraph==0.17.4
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.2
msoffcrypto-tool==5.4.2
olefile==0.47
packaging==25.0
pefile==2023.2.7
pycparser==2.23
pycryptodomex==3.23.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzipper==0.3.6
rarfile==4.2
setuptools==80.9.0
tqdm==4.67.1
